add SpeechApiUtils class
